move convert_to_dataset, convert_to_dataarray to type_handling
